Underpin REST client in shopify-api
