+++ .changeset/old-horses-change.md
@@ -1,0 +1,5 @@
+---
+"@shopify/shopify-api": minor
+---
+
+Underpinned the new REST client from `admin-api-client` into `shopify-api`, so apps can access it as a standalone client as well.
