+++ packages/shopify-api/adapters/__e2etests__/test_apps/test-web-api-app.ts
@@ -4,9 +4,10 @@ import {
   DataType,
   PutRequestParams,
   PostRequestParams,
-} from '../../../lib/clients/http_client/types';
-import {httpClientClass} from '../../../lib/clients/http_client/http_client';
-import {config, matchHeaders} from '../utils';
+} from '../../../lib/clients/types';
+import {Session} from '../../../lib';
+import {restClientClass} from '../../../lib/clients/admin';
+import {config, matchHeaders, session} from '../utils';
 import {
   TestResponse,
   TestConfig,
@@ -21,9 +22,12 @@ const RED = '\x1b[31m';
 const GREEN = '\x1b[32m';
 const RESET = '\x1b[39m';
 
-const HttpClient = httpClientClass(config, 'http');
+const RestClient = restClientClass({
+  config: {...config, hostScheme: 'http'},
+  formatPaths: false,
+});
 
-const defaultRetryTimer = HttpClient.RETRY_WAIT_TIME;
+const defaultRetryTimer = RestClient.RETRY_WAIT_TIME;
 let testCount = 0;
 
 function params(
@@ -37,8 +41,7 @@ function params(
 }
 
 function setRestClientRetryTime(time: number) {
-  // We de-type HttpClient here so we can alter its readonly time property
-  (HttpClient as any).RETRY_WAIT_TIME = time;
+  (RestClient as any).RETRY_WAIT_TIME = time;
 }
 
 /* eslint-disable-next-line import/no-anonymous-default-export */
@@ -49,7 +52,9 @@ export default {
       10,
     );
     const apiServer = `localhost:${apiServerPort}`;
-    const client = new HttpClient({domain: apiServer});
+    const client = new RestClient({
+      session: new Session({...session, shop: apiServer}),
+    });
 
     if (req.method === 'POST') {
       const testConfig: TestConfig = await req.json();
@@ -91,11 +96,11 @@ export default {
         if (typeof testRequest.retryTimeoutTimer !== 'undefined') {
           setRestClientRetryTime(testRequest.retryTimeoutTimer);
           console.log(
-            `[webApi] RETRY_TIME_WAIT (BEFORE) = ${HttpClient.RETRY_WAIT_TIME}\n`,
+            `[webApi] RETRY_TIME_WAIT (BEFORE) = ${RestClient.RETRY_WAIT_TIME}\n`,
           );
           if (testRequest.retryTimeoutTimer !== 0) {
             console.log(
-              `[webApi] setting setTimeout @ ${HttpClient.RETRY_WAIT_TIME} ms\n`,
+              `[webApi] setting setTimeout @ ${RestClient.RETRY_WAIT_TIME} ms\n`,
             );
 
             retryTimeout = setTimeout(() => {
@@ -105,7 +110,7 @@ export default {
                 );
               } catch (error) {
                 console.log(
-                  `[webApi] ${RED}setTimeout fired!${RESET} @ ${HttpClient.RETRY_WAIT_TIME}\n`,
+                  `[webApi] ${RED}setTimeout fired!${RESET} @ ${RestClient.RETRY_WAIT_TIME}\n`,
                 );
                 testFailedDebug = JSON.stringify({
                   errorMessageReceived: error.message,
