+++ packages/shopify-api/adapters/__e2etests__/test_suite.ts
@@ -1,4 +1,4 @@
-import {DataType} from '../../lib/clients/http_client/types';
+import {DataType} from '../../lib/clients/types';
 import ProcessedQuery from '../../lib/utils/processed-query';
 
 import {TestType, initTestRequest, initTestResponse} from './test_config_types';
@@ -193,6 +193,18 @@ export const testSuite = [
     },
   },
   {
+    name: 'gracefully handles 503 error',
+    config: {
+      testRequest: initTestRequest({url: '/url/path/503'}),
+      expectedResponse: initTestResponse({
+        statusCode: 503,
+        statusText: 'Did not work',
+        errorType: 'HttpInternalError',
+        expectRequestId: 'Request id header',
+      }),
+    },
+  },
+  {
     name: 'allows custom headers',
     config: {
       testRequest: initTestRequest({
@@ -225,17 +237,6 @@ export const testSuite = [
       }),
       expectedResponse: initTestResponse({
         headers: {'user-agent': 'My lowercase agent'},
-      }),
-    },
-  },
-  {
-    name: 'fails with invalid retry count',
-    config: {
-      testRequest: initTestRequest({tries: -1}),
-      expectedResponse: initTestResponse({
-        statusCode: 500,
-        statusText: 'Did not work',
-        errorType: 'HttpRequestError',
       }),
     },
   },
