+++ packages/shopify-api/lib/clients/admin/graphql/client.ts
@@ -9,15 +9,15 @@ import {
 
 import {ApiVersion} from '../../../types';
 import {ConfigInterface} from '../../../base-types';
-import {RequestReturn} from '../../http_client/types';
-import {Session} from '../../../session/session';
-import {logger} from '../../../logger';
-import * as ShopifyErrors from '../../../error';
 import type {
+  RequestReturn,
   GraphqlParams,
   GraphqlClientParams,
   GraphqlQueryOptions,
 } from '../../types';
+import {Session} from '../../../session/session';
+import {logger} from '../../../logger';
+import * as ShopifyErrors from '../../../error';
 import {abstractFetch} from '../../../../runtime';
 import {
   clientLoggerFactory,
