+++ packages/shopify-api/lib/clients/admin/rest/client.ts
@@ -1,10 +1,24 @@
-import {getHeader} from '../../../../runtime/http';
-import {ApiVersion, ShopifyHeader} from '../../../types';
+import {
+  AdminRestApiClient,
+  createAdminRestApiClient,
+} from '@shopify/admin-api-client';
+import {Method} from '@shopify/network';
+
+import {
+  clientLoggerFactory,
+  getUserAgent,
+  throwFailedRequest,
+} from '../../common';
+import {
+  HashFormat,
+  NormalizedRequest,
+  abstractFetch,
+  canonicalizeHeaders,
+  createSHA256HMAC,
+  getHeader,
+} from '../../../../runtime';
 import {ConfigInterface} from '../../../base-types';
-import {RequestParams} from '../../http_client/types';
 import * as ShopifyErrors from '../../../error';
-import {HttpClient} from '../../http_client/http_client';
-import {Session} from '../../../session/session';
 import {logger} from '../../../logger';
 import {
   RestRequestReturn,
@@ -12,60 +26,174 @@ import {
   RestClientParams,
   PageInfoParams,
 } from '../types';
+import type {
+  RequestParams,
+  GetRequestParams,
+  PutRequestParams,
+  PostRequestParams,
+  DeleteRequestParams,
+} from '../../types';
 
 export interface RestClientClassParams {
   config: ConfigInterface;
+  formatPaths?: boolean;
 }
 
-export class RestClient extends HttpClient {
+interface DeprecationInterface {
+  message: string | null;
+  path: string;
+  body?: string;
+}
+
+export class RestClient {
+  public static config: ConfigInterface;
+  public static formatPaths: boolean;
+
   static LINK_HEADER_REGEXP = /<([^<]+)>; rel="([^"]+)"/;
   static DEFAULT_LIMIT = '50';
+  static RETRY_WAIT_TIME = 1000;
 
-  public static config: ConfigInterface;
+  static readonly DEPRECATION_ALERT_DELAY = 300000;
+  loggedDeprecations: {[key: string]: number} = {};
 
-  readonly session: Session;
-  readonly apiVersion?: ApiVersion;
+  readonly client: AdminRestApiClient;
 
-  public constructor(params: RestClientParams) {
-    super({domain: params.session.shop});
-
+  public constructor({session, apiVersion}: RestClientParams) {
     const config = this.restClass().config;
 
-    if (!config.isCustomStoreApp && !params.session.accessToken) {
+    if (!config.isCustomStoreApp && !session.accessToken) {
       throw new ShopifyErrors.MissingRequiredArgument(
         'Missing access token when creating REST client',
       );
     }
 
-    if (params.apiVersion) {
+    if (apiVersion) {
       const message =
-        params.apiVersion === config.apiVersion
-          ? `REST client has a redundant API version override to the default ${params.apiVersion}`
-          : `REST client overriding default API version ${config.apiVersion} with ${params.apiVersion}`;
+        apiVersion === config.apiVersion
+          ? `REST client has a redundant API version override to the default ${apiVersion}`
+          : `REST client overriding default API version ${config.apiVersion} with ${apiVersion}`;
 
       logger(config).debug(message);
     }
 
-    this.session = params.session;
-    this.apiVersion = params.apiVersion;
+    const customStoreAppAccessToken =
+      config.adminApiAccessToken ?? config.apiSecretKey;
+
+    this.client = createAdminRestApiClient({
+      scheme: config.hostScheme,
+      storeDomain: session.shop,
+      apiVersion: apiVersion ?? config.apiVersion,
+      accessToken: config.isCustomStoreApp
+        ? customStoreAppAccessToken
+        : session.accessToken!,
+      customFetchApi: abstractFetch,
+      logger: clientLoggerFactory(config),
+      userAgentPrefix: getUserAgent(config),
+      defaultRetryTime: this.restClass().RETRY_WAIT_TIME,
+      formatPaths: this.restClass().formatPaths,
+    });
   }
 
-  protected async request<T = unknown>(
+  /**
+   * Performs a GET request on the given path.
+   */
+  public async get<T = any>(params: GetRequestParams) {
+    return this.request<T>({method: Method.Get, ...params});
+  }
+
+  /**
+   * Performs a POST request on the given path.
+   */
+  public async post<T = any>(params: PostRequestParams) {
+    return this.request<T>({method: Method.Post, ...params});
+  }
+
+  /**
+   * Performs a PUT request on the given path.
+   */
+  public async put<T = any>(params: PutRequestParams) {
+    return this.request<T>({method: Method.Put, ...params});
+  }
+
+  /**
+   * Performs a DELETE request on the given path.
+   */
+  public async delete<T = any>(params: DeleteRequestParams) {
+    return this.request<T>({method: Method.Delete, ...params});
+  }
+
+  private async request<T = any>(
     params: RequestParams,
   ): Promise<RestRequestReturn<T>> {
-    const customStoreAppAccessToken =
-      this.restClass().config.adminApiAccessToken ??
-      this.restClass().config.apiSecretKey;
-    params.extraHeaders = {
-      [ShopifyHeader.AccessToken]: this.restClass().config.isCustomStoreApp
-        ? customStoreAppAccessToken
-        : (this.session.accessToken as string),
-      ...params.extraHeaders,
+    const requestParams = {
+      headers: {
+        ...params.extraHeaders,
+        ...(params.type ? {'Content-Type': params.type.toString()} : {}),
+      },
+      retries: params.tries ? params.tries - 1 : undefined,
+      searchParams: params.query,
     };
 
-    const ret: RestRequestReturn<T> = await super.request<T>(params);
+    let response: Response;
+    switch (params.method) {
+      case Method.Get:
+        response = await this.client.get(params.path, requestParams);
+        break;
+      case Method.Put:
+        response = await this.client.put(params.path, {
+          ...requestParams,
+          data: params.data!,
+        });
+        break;
+      case Method.Post:
+        response = await this.client.post(params.path, {
+          ...requestParams,
+          data: params.data!,
+        });
+        break;
+      case Method.Delete:
+        response = await this.client.delete(params.path, requestParams);
+        break;
+      default:
+        throw new ShopifyErrors.InvalidRequestError(
+          `Unsupported request method '${params.method}'`,
+        );
+    }
 
-    const link = getHeader(ret.headers, 'link');
+    const body = await response.text();
+    const responseHeaders = canonicalizeHeaders(
+      Object.fromEntries(response.headers.entries()),
+    );
+
+    if (!response.ok) {
+      throwFailedRequest(
+        body,
+        {
+          statusCode: response.status,
+          statusText: response.statusText,
+          headers: responseHeaders,
+        },
+        (params.tries ?? 1) <= 1,
+      );
+    }
+
+    const jsonBody = JSON.parse(body);
+    const requestReturn: RestRequestReturn<T> = {
+      body: jsonBody,
+      headers: responseHeaders,
+    };
+
+    await this.logDeprecations(
+      {
+        method: params.method,
+        url: params.path,
+        headers: requestParams.headers,
+        body: params.data ? JSON.stringify(params.data) : undefined,
+      },
+      requestReturn,
+    );
+
+    const link = response.headers.get('Link');
     if (link !== undefined) {
       const pageInfo: PageInfo = {
         limit: params.query?.limit
@@ -106,21 +234,10 @@ export class RestClient extends HttpClient {
         }
       }
 
-      ret.pageInfo = pageInfo;
+      requestReturn.pageInfo = pageInfo;
     }
 
-    return ret;
-  }
-
-  protected getRequestPath(path: string): string {
-    const cleanPath = super.getRequestPath(path);
-    if (cleanPath.startsWith('/admin')) {
-      return `${cleanPath.replace(/\.json$/, '')}.json`;
-    } else {
-      return `/admin/api/${
-        this.apiVersion || this.restClass().config.apiVersion
-      }${cleanPath.replace(/\.json$/, '')}.json`;
-    }
+    return requestReturn;
   }
 
   private restClass() {
@@ -137,16 +254,59 @@ export class RestClient extends HttpClient {
       query: Object.fromEntries(url.searchParams.entries()),
     };
   }
+
+  private async logDeprecations(
+    request: NormalizedRequest,
+    response: RestRequestReturn,
+  ) {
+    const config = this.restClass().config;
+
+    const deprecationReason = getHeader(
+      response.headers,
+      'X-Shopify-API-Deprecated-Reason',
+    );
+    if (deprecationReason) {
+      const deprecation: DeprecationInterface = {
+        message: deprecationReason,
+        path: request.url,
+      };
+
+      if (request.body) {
+        // This can only be a string, since we're always converting the body before calling this method
+        deprecation.body = `${(request.body as string).substring(0, 100)}...`;
+      }
+
+      const depHash = await createSHA256HMAC(
+        config.apiSecretKey,
+        JSON.stringify(deprecation),
+        HashFormat.Hex,
+      );
+
+      if (
+        !Object.keys(this.loggedDeprecations).includes(depHash) ||
+        Date.now() - this.loggedDeprecations[depHash] >=
+          RestClient.DEPRECATION_ALERT_DELAY
+      ) {
+        this.loggedDeprecations[depHash] = Date.now();
+
+        const stack = new Error().stack;
+        const message = `API Deprecation Notice ${new Date().toLocaleString()} : ${JSON.stringify(
+          deprecation,
+        )}  -  Stack Trace: ${stack}`;
+        await logger(config).warning(message);
+      }
+    }
+  }
 }
 
 export function restClientClass(
   params: RestClientClassParams,
 ): typeof RestClient {
-  const {config} = params;
+  const {config, formatPaths} = params;
 
   class NewRestClient extends RestClient {
     public static config = config;
-    public static scheme = 'https';
+    public static formatPaths = formatPaths === undefined ? true : formatPaths;
   }
 
   Reflect.defineProperty(NewRestClient, 'name', {
