+++ packages/shopify-api/lib/clients/types.ts
@@ -1,18 +1,19 @@
 import {
   AllOperations,
   ApiClientRequestOptions,
+  SearchParams,
 } from '@shopify/admin-api-client';
+import {Method} from '@shopify/network';
 
 import {Session} from '../session/session';
-import {ApiVersion} from '../types';
+import type {ApiVersion} from '../types';
+import {Headers} from '../../runtime/http';
 
 import {GraphqlClient} from './admin/graphql/client';
 import {StorefrontClient} from './storefront/client';
-import {GraphqlProxy} from './graphql_proxy/types';
+import type {GraphqlProxy} from './graphql_proxy/types';
 import {RestClient} from './admin/rest/client';
-import {PostRequestParams} from './http_client/types';
 
-export * from './http_client/types';
 export * from './admin/types';
 export * from './graphql_proxy/types';
 
@@ -20,6 +21,44 @@ export interface ClientArgs {
   session: Session;
   apiVersion?: ApiVersion;
   retries?: number;
+}
+
+export interface HeaderParams {
+  [key: string]: string | number | string[];
+}
+
+/* eslint-disable @shopify/typescript/prefer-pascal-case-enums */
+export enum DataType {
+  JSON = 'application/json',
+  GraphQL = 'application/graphql',
+  URLEncoded = 'application/x-www-form-urlencoded',
+}
+/* eslint-enable @shopify/typescript/prefer-pascal-case-enums */
+
+export interface GetRequestParams {
+  path: string;
+  type?: DataType;
+  data?: {[key: string]: unknown} | string;
+  query?: SearchParams;
+  extraHeaders?: HeaderParams;
+  tries?: number;
+}
+
+export type PostRequestParams = GetRequestParams & {
+  data: {[key: string]: unknown} | string;
+};
+
+export type PutRequestParams = PostRequestParams;
+
+export type DeleteRequestParams = GetRequestParams;
+
+export type RequestParams = (GetRequestParams | PostRequestParams) & {
+  method: Method;
+};
+
+export interface RequestReturn<T = unknown> {
+  body: T;
+  headers: Headers;
 }
 
 export type GraphqlParams = Omit<PostRequestParams, 'path' | 'type'>;
