+++ packages/shopify-api/rest/base.ts
@@ -1,7 +1,7 @@
 import {RestResourceError} from '../lib/error';
 import {Session} from '../lib/session/session';
 import {PageInfo, RestRequestReturn} from '../lib/clients/admin/types';
-import {DataType} from '../lib/clients/http_client/types';
+import {DataType} from '../lib/clients/types';
 import {RestClient} from '../lib/clients/admin/rest/client';
 import {ApiVersion} from '../lib/types';
 import {ConfigInterface} from '../lib/base-types';
